+++ server/api/admin/products.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { Product, User } = require('../../db/models');
+const { Product } = require('../../db/models');
 
 module.exports = router
